+++ src/midend/MFB/lib/sage/namespace-declaration.cpp
@@ -47,7 +47,7 @@ bool Driver<Sage>::resolveValidParent<SgNamespaceSymbol>(SgNamespaceSymbol * sym
 }
 
 template <>
-void Driver<Sage>::loadSymbols<SgNamespaceDeclarationStatement>(unsigned file_id, SgSourceFile * file) {
+void Driver<Sage>::loadSymbols<SgNamespaceDeclarationStatement>(size_t file_id, SgSourceFile * file) {
   std::vector<SgNamespaceDeclarationStatement *> namespace_decl = SageInterface::querySubTree<SgNamespaceDeclarationStatement>(file);
 
   std::set<SgNamespaceSymbol *> namespace_symbols;
@@ -70,7 +70,7 @@ void Driver<Sage>::loadSymbols<SgNamespaceDeclarationStatement>(unsigned file_id
     }
 }
 
-Sage<SgNamespaceDeclarationStatement>::object_desc_t::object_desc_t(const std::string & name_, SgNamespaceSymbol * parent_, unsigned file_id_) :
+Sage<SgNamespaceDeclarationStatement>::object_desc_t::object_desc_t(const std::string & name_, SgNamespaceSymbol * parent_, size_t file_id_) :
   name(name_),
   parent(parent_),
   file_id(file_id_)
@@ -127,7 +127,7 @@ Sage<SgNamespaceDeclarationStatement>::build_scopes_t Driver<Sage>::getBuildScop
 
   assert(desc.file_id != 0);
 
-  std::map<unsigned, SgSourceFile *>::iterator it_file = id_to_file_map.find(desc.file_id);
+  std::map<size_t, SgSourceFile *>::iterator it_file = id_to_file_map.find(desc.file_id);
   assert(it_file != id_to_file_map.end());
   SgSourceFile * file = it_file->second;
   assert(file != NULL);
